+++ fltk-derive/src/image.rs
@@ -30,18 +30,16 @@ pub fn impl_image_trait(ast: &DeriveInput) -> TokenStream {
         impl Clone for #name {
             fn clone(&self) -> Self {
                 assert!(!self.was_deleted());
-                let mut x = self._refcount.lock().unwrap();
-                *x += 1;
-                #name { _inner: self._inner, _refcount: Arc::from(Mutex::from(*x)) }
+                let x = self._refcount.fetch_add(1, Ordering::Relaxed);
+                #name { _inner: self._inner, _refcount: AtomicUsize::new(x) }
             }
         }
 
         impl Drop for #name {
             fn drop(&mut self) {
                 if !self.was_deleted() {
-                    let mut x = self._refcount.lock().unwrap();
-                    *x -= 1;
-                    if *x == 0 {
+                    let x = self._refcount.fetch_sub(1, Ordering::Relaxed);
+                    if x == 0 {
                         unsafe {
                             #delete(self._inner);
                             self._inner = std::ptr::null_mut();
@@ -59,7 +57,7 @@ pub fn impl_image_trait(ast: &DeriveInput) -> TokenStream {
                     assert!(!img.is_null());
                     #name {
                         _inner: img,
-                        _refcount: Arc::from(Mutex::from(1))
+                        _refcount: AtomicUsize::new(1)
                     }
                 }
             }
@@ -95,7 +93,7 @@ pub fn impl_image_trait(ast: &DeriveInput) -> TokenStream {
                     assert!(!ptr.is_null());
                     #name {
                         _inner: ptr as *mut #ptr_name,
-                        _refcount: Arc::from(Mutex::from(2))
+                        _refcount: AtomicUsize::new(2),
                     }
                 }
             }
@@ -211,7 +209,7 @@ pub fn impl_image_trait(ast: &DeriveInput) -> TokenStream {
 
             unsafe fn increment_arc(&mut self) {
                 assert!(!self.was_deleted());
-                *self._refcount.lock().unwrap() += 1;
+                self._refcount.fetch_add(1, Ordering::Relaxed);
             }
 
             fn was_deleted(&self) -> bool {