+++ fltk/src/text.rs
@@ -5,14 +5,14 @@ use std::{
     ffi::{CStr, CString},
     mem,
     os::raw,
-    sync::{Arc, Mutex},
+    sync::atomic::{AtomicUsize, Ordering},
 };
 
 /// Wraps a text buffer, Cloning a text buffer invalidates the underlying pointer, thus the no derive(Clone)
 #[derive(Debug)]
 pub struct TextBuffer {
     _inner: *mut Fl_Text_Buffer,
-    _refcount: Arc<Mutex<usize>>,
+    _refcount: AtomicUsize,
 }
 
 impl TextBuffer {
@@ -23,7 +23,7 @@ impl TextBuffer {
             assert!(!text_buffer.is_null());
             TextBuffer {
                 _inner: text_buffer,
-                _refcount: Arc::from(Mutex::from(1)),
+                _refcount: AtomicUsize::new(1),
             }
         }
     }
@@ -51,7 +51,7 @@ impl TextBuffer {
         assert!(!ptr.is_null());
         TextBuffer {
             _inner: ptr,
-            _refcount: Arc::from(Mutex::from(2)),
+            _refcount: AtomicUsize::new(2),
         }
     }
 
@@ -549,11 +549,10 @@ unsafe impl Send for TextBuffer {}
 impl Clone for TextBuffer {
     fn clone(&self) -> TextBuffer {
         assert!(!self._inner.is_null());
-        let mut x = self._refcount.lock().unwrap();
-        *x += 1;
+        let x = self._refcount.fetch_add(1, Ordering::Relaxed);
         TextBuffer {
             _inner: self._inner,
-            _refcount: Arc::from(Mutex::from(*x)),
+            _refcount: AtomicUsize::new(x),
         }
     }
 }
@@ -561,9 +560,8 @@ impl Clone for TextBuffer {
 impl Drop for TextBuffer {
     fn drop(&mut self) {
         assert!(!self._inner.is_null());
-        let mut x = self._refcount.lock().unwrap();
-        *x -= 1;
-        if *x == 0 {
+        let x = self._refcount.fetch_sub(1, Ordering::Relaxed);
+        if x == 0 {
             unsafe {
                 Fl_Text_Buffer_delete(self._inner);
             }