+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.10.3] - Unreleased
 ### Changes
 - Change return type of Printer::begin_job().
+- Use AtomicUsize for refcounting instead of Arc<Mutex>.
 
 ## [0.10.3] - 2020-11-04
 ### Changes